Add files via upload
The project is intended to maintain track of two key bodies: a library's administrator and a library's user. This project keeps track of all of the books in the library, whether they are issued or returned, as well as all of the library's members. If a new book is added to the library, the administrator can insert it, or delete it if it.